fix(app.js): fixed sql query

+++ app/dbinit.sql
@@ -3,9 +3,9 @@ CREATE DATABASE erste;
 USE erste;
 
 CREATE TABLE bankcards (
-	card_id INT(7) NOT NULL AUTO_INCREMENT,
+	card_id INT NOT NULL AUTO_INCREMENT,
 	card_type ENUM('Mastercard Standard', 'Maestro', 'Mastercard Gold', 'Mastercard World Gold', 'Mastercard Standard Devisa', 'Maestro Student', 'Visa Classic', 'Visa Virtual', 'Visa Electron' ) NOT NULL,
-	card_num INT(16) NOT NULL,
+	card_num BIGINT NOT NULL,
 	card_valid VARCHAR(5) NOT NULL,
 	card_owner VARCHAR(40) NOT NULL,
 	card_hash VARCHAR(150) NOT NULL,
@@ -14,8 +14,8 @@ CREATE TABLE bankcards (
 );
 
 CREATE TABLE contact (
-	contact_id INT(7) NOT NULL AUTO_INCREMENT,
-	card_id INT(7) NOT NULL,
+	contact_id INT NOT NULL AUTO_INCREMENT,
+	card_id INT NOT NULL,
 	contact_type VARCHAR(20) NOT NULL,
 	contact_data VARCHAR(40) NOT NULL,
 	PRIMARY KEY (contact_id),
